# Faster gaussian error function