--- a/perf/gfoidl.Stochastics.Benchmarks/gfoidl.Stochastics.Benchmarks.csproj
+++ b/perf/gfoidl.Stochastics.Benchmarks/gfoidl.Stochastics.Benchmarks.csproj
@@ -25,6 +25,10 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <Content Include="..\..\native-out\*" Visible="false" CopyToOutputDirectory="PreserveNewest" />
+    </ItemGroup>
+
+    <ItemGroup>
         <PackageReference Include="BenchmarkDotNet" Version="0.10.11" />
         <PackageReference Include="System.Memory" Version="4.4.0-preview2-25405-01" />
         <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.4.0" />