--- a/source/gfoidl.Stochastics.Native-linux-64/build.sh
+++ b/source/gfoidl.Stochastics.Native-linux-64/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+clang++ -o libgfoidl-Stochastics-Native.so -Ofast -std=c++11 -shared -fPIC special_functions.cpp