--- a/source/gfoidl.Stochastics.Native-linux-64/special_functions.cpp
+++ b/source/gfoidl.Stochastics.Native-linux-64/special_functions.cpp
@@ -1,0 +1,24 @@
+#include <cmath>
+#include "special_functions.h"
+//-----------------------------------------------------------------------------
+double gaussian_error_function(double x)
+{
+    return std::erf(x);
+}
+//-----------------------------------------------------------------------------
+double gaussian_error_function_complementary(double x)
+{
+    return std::erfc(x);
+}
+//-----------------------------------------------------------------------------
+void gaussian_error_function_vector(double * values, double * result, const int n)
+{
+    for (int i = 0; i < n; ++i)
+        result[i] = erf(values[i]);
+}
+//-----------------------------------------------------------------------------
+void gaussian_error_function_complementary_vector(double * values, double * result, const int n)
+{
+    for (int i = 0; i < n; ++i)
+        result[i] = erfc(values[i]);
+}