--- a/source/gfoidl.Stochastics.Native-linux-64/special_functions.cpp
+++ b/source/gfoidl.Stochastics.Native-linux-64/special_functions.cpp
@@ -11,13 +11,13 @@ double gaussian_error_function_complementary(double x)
     return std::erfc(x);
 }
 //-----------------------------------------------------------------------------
-void gaussian_error_function_vector(double * values, double * result, const int n)
+void gaussian_error_function_vector(double* values, double* result, const int n)
 {
     for (int i = 0; i < n; ++i)
         result[i] = erf(values[i]);
 }
 //-----------------------------------------------------------------------------
-void gaussian_error_function_complementary_vector(double * values, double * result, const int n)
+void gaussian_error_function_complementary_vector(double* values, double* result, const int n)
 {
     for (int i = 0; i < n; ++i)
         result[i] = erfc(values[i]);