--- a/source/gfoidl.Stochastics.Native-win-x64/dll.h
+++ b/source/gfoidl.Stochastics.Native-win-x64/dll.h
@@ -1,0 +1,11 @@
+// The following ifdef block is the standard way of creating macros which make exporting 
+// from a DLL simpler. All files within this DLL are compiled with the GFOIDLSTOCHASTICSNATIVE_EXPORTS
+// symbol defined on the command line. This symbol should not be defined on any project
+// that uses this DLL. This way any other project whose source files include this file see 
+// DLL_API functions as being imported from a DLL, whereas this DLL sees symbols
+// defined with this macro as being exported.
+#ifdef GFOIDLSTOCHASTICSNATIVE_EXPORTS
+#define DLL_API __declspec(dllexport)
+#else
+#define DLL_API __declspec(dllimport)
+#endif