--- a/tests/gfoidl.Stochastics.Tests/gfoidl.Stochastics.Tests.csproj
+++ b/tests/gfoidl.Stochastics.Tests/gfoidl.Stochastics.Tests.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <TargetFramework>netcoreapp2.0</TargetFramework>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     </PropertyGroup>
 
     <ItemGroup>