… native-out/gfoidl-Stochastics-Native.dll
… native-out/libgfoidl-Stochastics-Native.so
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/linux-x64/report-github.md
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/linux-x64/report.csv
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/linux-x64/report.html
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/win-x64/report-github.md
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/win-x64/report.csv
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionBenchmarks/win-x64/report.html
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/linux-x64/report.csv
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/linux-x64/report.html
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/linux-x64/report.md
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/win-x64/report.csv
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/win-x64/report.html
… perf/gfoidl.Stochastics.Benchmarks/BenchmarkDotNet.Artifacts/results/ErrorFunctionVectorSizeBenchmarks/win-x64/report.md
… perf/gfoidl.Stochastics.Benchmarks/ErrorFunctionVectorSizeBenchmarks.cs
… perf/gfoidl.Stochastics.Benchmarks/Properties/launchSettings.json
… source/gfoidl.Stochastics.Native-win-x64/gfoidl.Stochastics.Native-win-x64.filters
… source/gfoidl.Stochastics.Native-win-x64/gfoidl.Stochastics.Native-win-x64.vcxproj
… source/gfoidl.Stochastics.Native-win-x64/special_functions.cpp
… source/gfoidl.Stochastics.Native-win-x64/special_functions.h
… source/gfoidl.Stochastics.Native-win-x64/stdafx.cpp
… source/gfoidl.Stochastics.Native-win-x64/stdafx.h
… source/gfoidl.Stochastics.Native-win-x64/targetver.h
… source/gfoidl.Stochastics/Native/NativeMethods.cs
… source/gfoidl.Stochastics/SpecialFunctions.cs
… source/gfoidl.Stochastics/Statistics/OutlierDetection/ChauvenetOutlierDetection.cs
… source/gfoidl.Stochastics/gfoidl.Stochastics.csproj
… tests/gfoidl.Stochastics.Tests/SpecialFunctionsTests/ErrorFunction.cs