# Background / Motivation

Especially on Linux the `erf` is slow. See benchmarks below.

This is not a problem of .net! I don't know the exact root cause for this -- see https://github.com/gfoidl/erf_linux_problem

# Description

For _win-64_ (vc++) and _linux-x64_ (clang -Ofast) native methods are implemented. On behalf to not make the p/invoke chatty, new methods that operate on vectors got added, so that it is chunky :wink:

Caching was tried, but rejected by design due the assumption that datapoints won't match exactly (GetHashCode). Furthermore only on linux-x64 was a slight perf-win. On win-x64 it was slower. Other platforms not tested.

# Benchmarks

## win-x64

``` ini

BenchmarkDotNet=v0.10.11, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Processor=Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), ProcessorCount=8
Frequency=2742191 Hz, Resolution=364.6719 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|        Method |     Mean |    Error |   StdDev | Scaled | ScaledSD |
|-------------- |---------:|---------:|---------:|-------:|---------:|
|          Erf1 | 151.0 ns | 3.031 ns | 3.491 ns |   1.00 |     0.00 |
|          Erf2 | 140.4 ns | 2.583 ns | 4.386 ns |   0.93 |     0.04 |
|        ErfCpp | 146.3 ns | 2.345 ns | 2.194 ns |   0.97 |     0.03 |
| ErfCpp_Vector | 109.6 ns | 2.147 ns | 2.556 ns |   0.73 |     0.02 |

## linux-x64

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), ProcessorCount=4
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT


```
|        Method |      Mean |    Error |   StdDev | Scaled | ScaledSD |
|-------------- |----------:|---------:|---------:|-------:|---------:|
|          Erf1 | 251.88 ns | 4.964 ns | 4.875 ns |   1.00 |     0.00 |
|          Erf2 | 147.97 ns | 2.966 ns | 5.715 ns |   0.59 |     0.02 |
|        ErfCpp |  81.19 ns | 1.882 ns | 3.844 ns |   0.32 |     0.02 |
| ErfCpp_Vector |  81.95 ns | 1.501 ns | 1.331 ns |   0.33 |     0.01 |

## Notes

On win-x64 for _erf_ and _erfc_ the C# implementation is used, 'cause it's faster.

# Prospects

Will show effect in https://github.com/gfoidl/Stochastics/issues/31